Fix attribution of HTML attributes.
In the scenario of having extra HTML attributes without some content being
passed in, the attributes hash was being set to nil while the content variable
holding the attributes hash.

In the scenario:

```ruby
form_for :book do
  text_area :description, rows: 5
end
```

The `content` argument was `{ rows: 5 }` and `attributes` was `{}`.

This is still a pretty ugly solution. I need some feedback on this.